big bad commit of matt's stuff
