+++ porespy/export/__init__.py
@@ -1,1 +1,2 @@
-from .__funcs__ import vox2vtk
+from . import evtk
+from .__funcs__ import vox2vtk, im2vtk
